+++ src/app/data_harvester/processes/macos.rs
@@ -7,13 +7,13 @@ use crate::{data_harvester::processes::UserTable, Pid};
 mod sysctl_bindings;
 
 pub fn get_process_data(
-    sys: &System, use_current_cpu_total: bool, per_core_percentage: bool, mem_total_kb: u64,
+    sys: &System, use_current_cpu_total: bool, unnormalized_cpu: bool, mem_total_kb: u64,
     user_table: &mut UserTable,
 ) -> crate::utils::error::Result<Vec<ProcessHarvest>> {
     super::macos_freebsd::get_process_data(
         sys,
         use_current_cpu_total,
-        per_core_percentage,
+        unnormalized_cpu,
         mem_total_kb,
         user_table,
         get_macos_process_cpu_usage,